+++ src/runtime/StolenReference.cs
@@ -1,18 +1,46 @@
 namespace Python.Runtime
 {
     using System;
+    using System.Diagnostics.Contracts;
 
     /// <summary>
-    /// Should only be used for the arguments of Python C API functions, that steal references
+    /// Should only be used for the arguments of Python C API functions, that steal references,
+    /// and internal <see cref="PyObject"/> constructors.
     /// </summary>
     [NonCopyable]
     readonly ref struct StolenReference
     {
-        readonly IntPtr pointer;
+        internal readonly IntPtr Pointer;
 
         internal StolenReference(IntPtr pointer)
         {
-            this.pointer = pointer;
+            Pointer = pointer;
         }
+
+        [Pure]
+        public static bool operator ==(in StolenReference reference, NullOnly @null)
+            => reference.Pointer == IntPtr.Zero;
+        [Pure]
+        public static bool operator !=(in StolenReference reference, NullOnly @null)
+            => reference.Pointer != IntPtr.Zero;
+
+        [Pure]
+        public override bool Equals(object obj)
+        {
+            if (obj is IntPtr ptr)
+                return ptr == Pointer;
+
+            return false;
+        }
+
+        [Pure]
+        public override int GetHashCode() => Pointer.GetHashCode();
+    }
+
+    static class StolenReferenceExtensions
+    {
+        [Pure]
+        public static IntPtr DangerousGetAddressOrNull(this in StolenReference reference)
+            => reference.Pointer;
     }
 }