+++ src/runtime/typemanager.cs
@@ -177,7 +177,7 @@ namespace Python.Runtime
 
             if (Runtime.PyType_Ready(type) != 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
 
             var dict = new BorrowedReference(Marshal.ReadIntPtr(type, TypeOffset.tp_dict));
@@ -300,7 +300,7 @@ namespace Python.Runtime
 
             if (Runtime.PyType_Ready(type) != 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
 
             var dict = new BorrowedReference(Marshal.ReadIntPtr(type, TypeOffset.tp_dict));
@@ -472,7 +472,7 @@ namespace Python.Runtime
 
             if (Runtime.PyType_Ready(type) != 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
 
             IntPtr dict = Marshal.ReadIntPtr(type, TypeOffset.tp_dict);
@@ -577,7 +577,7 @@ namespace Python.Runtime
 
             if (Runtime.PyType_Ready(type) != 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
 
             IntPtr tp_dict = Marshal.ReadIntPtr(type, TypeOffset.tp_dict);
@@ -926,14 +926,14 @@ namespace Python.Runtime
             if (Runtime.PyDict_SetItemString(globals, "__builtins__", Runtime.PyEval_GetBuiltins()) != 0)
             {
                 globals.Dispose();
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             const string code = "class A(object): pass";
             using var resRef = Runtime.PyRun_String(code, RunFlagType.File, globals, globals);
             if (resRef.IsNull())
             {
                 globals.Dispose();
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             resRef.Dispose();
             BorrowedReference A = Runtime.PyDict_GetItemString(globals, "A");